pyproject.toml
[project]
name = "mcpdemo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-groq>=0.3.2",
    "langchain-openai>=0.3.14",
    "mcp-use>=1.2.7",
    "python-dotenv>=1.1.0",
]
